requirements.txt
pandas==3.0.2
plotly==6.7.0
python-dotenv==1.2.2
requests==2.33.1
streamlit==1.56.0